pyproject.toml
[project]
name = "bear-mcp-python"
version = "0.1.0"
description = "MCP Server for Bear Notes with semantic search and AI summarization"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # MCP Server Framework
    "mcp>=0.4.0",
    
    # Configuration and validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    
    # Database (sqlite3 is built-in to Python)
    
    # Semantic analysis and AI
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    
    # AI Integration
    "httpx>=0.24.0",  # For Ollama HTTP client
    "aiohttp>=3.8.0",
    
    # File monitoring
    "watchdog>=3.0.0",
    
    # Text processing
    "markdown>=3.5.0",
    "beautifulsoup4>=4.12.0",
    
    # Logging
    "structlog>=23.1.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    
    # Code quality
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
    
    # Type stubs
    "types-PyYAML>=6.0.0",
    "types-requests>=2.31.0",
]

[project.scripts]
bear-mcp = "bear_mcp.main:main"
bear-mcp-cli = "bear_mcp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
